// front/src/app/Models/paymentUrl.ts
export class paymentUrl {
  showId: string;
  movieId: string;
  theatreId: string;
  seats: string[];
  totalseat: number;
  showtime: string;
  date: string;
  showtimedate: string;

  //  constructor(showId: String,
  //     movieId: String, theatreId: String, seats: String[],
  //     totalseat: Number, showtime: String, date: String)
  //         {

  //             this.showId = showId;
  //             this.movieId = movieId;
  //             this.theatreId = theatreId;
  //             this.seats = seats;
  //             this.totalseat = totalseat;
  //             this.showtime = showtime;
  //             this.date = date;
  //     }
}